+++ nucleus/annotation_uploader.py
@@ -1,4 +1,5 @@
 import json
+from collections import Counter
 from typing import TYPE_CHECKING, Iterable, List, Optional, Sequence
 
 from nucleus.annotation import Annotation, SegmentationAnnotation
@@ -8,6 +9,7 @@ from nucleus.async_utils import (
     make_many_form_data_requests_concurrently,
 )
 from nucleus.constants import MASK_TYPE, SERIALIZED_REQUEST_KEY
+from nucleus.errors import DuplicateIDError
 from nucleus.payload_constructor import (
     construct_annotation_payload,
     construct_segmentation_payload,
@@ -207,6 +209,26 @@ class AnnotationUploader:
             return form_data, file_pointers
 
         return fn
+
+    @staticmethod
+    def check_for_duplicate_ids(annotations: Iterable[Annotation]):
+        """Do not allow annotations to have the same (annotation_id, reference_id) tuple"""
+
+        # some annotations like CategoryAnnotation do not have annotation_id attribute, and as such, we allow duplicates
+        tuple_ids = [
+            (ann.reference_id, ann.annotation_id)  # type: ignore
+            for ann in annotations
+            if hasattr(ann, "annotation_id")
+        ]
+        tuple_count = Counter(tuple_ids)
+        duplicates = {key for key, value in tuple_count.items() if value > 1}
+        if len(duplicates) > 0:
+            raise DuplicateIDError(
+                f"Duplicate annotations with the same (reference_id, annotation_id) properties found.\n"
+                f"Duplicates: {duplicates}\n"
+                f"To fix this, avoid duplicate annotations, or specify a different annotation_id attribute "
+                f"for the failing items."
+            )
 
 
 class PredictionUploader(AnnotationUploader):