+++ nucleus/metrics/__init__.py
@@ -15,3 +15,11 @@ from .polygon_metrics import (
     PolygonPrecision,
     PolygonRecall,
 )
+from .segmentation_metrics import (
+    SegmentationMaskToPolyMetric,
+    SegmentationToPolyAveragePrecision,
+    SegmentationToPolyIOU,
+    SegmentationToPolyMAP,
+    SegmentationToPolyPrecision,
+    SegmentationToPolyRecall,
+)