+++ CHANGELOG.md
@@ -5,11 +5,63 @@ All notable changes to the [Nucleus Python Client](https://github.com/scaleapi/n
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.10.4](https://github.com/scaleapi/nucleus-python-client/releases/tag/v0.10.4) - 2022-05-19
+## [0.11.2](https://github.com/scaleapi/nucleus-python-client/releases/tag/v0.11.2) - 2022-05-20
+
+### Changed
+
+- Restored backward compatibility of video constructor by adding back deprecated attachment_type argument
+
+## [0.11.1](https://github.com/scaleapi/nucleus-python-client/releases/tag/v0.11.1) - 2022-05-19
 
 ### Added
 
 - Exporting model predictions from a slice
+
+## [0.11.0](https://github.com/scaleapi/nucleus-python-client/releases/tag/v0.11.0) - 2022-05-13
+
+### Added
+
+- Segmentation prediction masks can now be evaluated against polygon annotation with new Validate functions
+- New function SegmentationToPolyIOU, configurable through client.validate.eval_functions.segmentation_to_poly_iou
+- New function SegmentationToPolyRecall, configurable through client.validate.eval_functions.segmentation_to_poly_recall
+- New function SegmentationToPolyPrecision, configurable through client.validate.eval_functions.segmentation_to_poly_precision
+- New function SegmentationToPolyMAP, configurable through client.validate.eval_functions.segmentation_to_poly_map
+- New function SegmentationToPolyAveragePrecision, configurable through client.validate.eval_functions.segmentation_to_poly_ap
+
+## [0.10.8](https://github.com/scaleapi/nucleus-python-client/releases/tag/v0.10.8) - 2022-05-10
+
+### Fixed
+
+- Add checks for duplicate (`reference_id`, `annotation_id`) when uploading Annotations or Predictions
+
+## [0.10.7](https://github.com/scaleapi/nucleus-python-client/releases/tag/v0.10.7) - 2022-05-09
+
+### Fixed
+
+- Add checks for duplicate reference IDs
+
+## [0.10.6](https://github.com/scaleapi/nucleus-python-client/releases/tag/v0.10.6) - 2022-05-06
+
+### Added
+
+- Video privacy mode
+
+### Changed
+
+- Removed attachment_type argument in video upload API
+
+## [0.10.5](https://github.com/scaleapi/nucleus-python-client/releases/tag/v0.10.5) - 2022-05-04
+
+### Fixed
+
+- Invalid polygons are dropped from PolygonMetric iou matching
+
+## [0.10.4](https://github.com/scaleapi/nucleus-python-client/releases/tag/v0.10.4)) - 2022-05-02
+
+### Added
+
+- Additional check added for KeypointsAnnotation names validation
+- MP4 video upload
 
 ## [0.10.3](https://github.com/scaleapi/nucleus-python-client/releases/tag/v0.10.3) - 2022-04-22